requirements.txt
pdfplumber==0.7.7
PyMuPDF==1.23.1
Pillow==10.1.0
regex==2024.7.14
opencv-python-headless==4.8.1.78
tqdm==4.66.1
